Revert "Created a package for the Component library"
