+++ Ashampoo-Translation-Systems.sln
@@ -22,7 +22,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ashampoo.Translation.System
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ashampoo.Translation.Systems.Formats.TsProj.Tests", "src\Ashampoo.Translation.Systems.Formats.TsProj\tests\Ashampoo.Translation.Systems.Formats.TsProj.Tests.csproj", "{4DDC2740-501A-4760-8384-21B9D2F994DD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ashampoo.Translation.Systems.Components", "src\Ashampoo.Translation.Systems.Components\src\Ashampoo.Translation.Systems.Components.csproj", "{D8C6BB19-A1D1-4CEC-A605-31B8A2C58FDC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ashampoo.Translation.Systems.Logging", "src\Ashampoo.Translation.Systems.Logging\src\Ashampoo.Translation.Systems.Logging.csproj", "{F2CA0A68-AE97-4BFF-9E5D-6902DF77DF57}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ashampoo.Translation.Systems.Logging.Tests", "src\Ashampoo.Translation.Systems.Logging\tests\Ashampoo.Translation.Systems.Logging.Tests.csproj", "{B3E8B243-573B-481B-AAD0-9825F9077BA9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ashampoo.Translation.Systems.Tools.ComponentLibrary", "src\Ashampoo.Translation.Systems.Tools.ComponentLibrary\Ashampoo.Translation.Systems.Tools.ComponentLibrary.csproj", "{D8C6BB19-A1D1-4CEC-A605-31B8A2C58FDC}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ashampoo.Translation.Systems.Formats.Abstractions.Tests", "src\Ashampoo.Translation.Systems.Formats.Abstractions\tests\Ashampoo.Translation.Systems.Formats.Abstractions.Tests.csproj", "{5C56057B-2927-4CB0-B9E2-13154699CD8E}"
 EndProject
@@ -88,6 +92,14 @@ Global
 		{4DDC2740-501A-4760-8384-21B9D2F994DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4DDC2740-501A-4760-8384-21B9D2F994DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4DDC2740-501A-4760-8384-21B9D2F994DD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F2CA0A68-AE97-4BFF-9E5D-6902DF77DF57}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F2CA0A68-AE97-4BFF-9E5D-6902DF77DF57}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F2CA0A68-AE97-4BFF-9E5D-6902DF77DF57}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F2CA0A68-AE97-4BFF-9E5D-6902DF77DF57}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B3E8B243-573B-481B-AAD0-9825F9077BA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B3E8B243-573B-481B-AAD0-9825F9077BA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B3E8B243-573B-481B-AAD0-9825F9077BA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B3E8B243-573B-481B-AAD0-9825F9077BA9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D8C6BB19-A1D1-4CEC-A605-31B8A2C58FDC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D8C6BB19-A1D1-4CEC-A605-31B8A2C58FDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D8C6BB19-A1D1-4CEC-A605-31B8A2C58FDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
