--- src/Ashampoo.Translation.Systems.Components/src/Notifications/FormatUploadedNotification.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-
-namespace Ashampoo.Translation.Systems.Components.Notifications;
-
-/// <summary>
-/// Notification for when a format was uploaded.
-/// </summary>
-public record FormatUploadedNotification : INotification;
