--- src/Ashampoo.Translation.Systems.Components/src/Notifications/SimpleWarning.cs
@@ -1,11 +1,0 @@
-﻿using MediatR;
-
-namespace Ashampoo.Translation.Systems.Components.Notifications;
-
-/// <summary>
-/// Notification to display a simple warning message.
-/// </summary>
-/// <param name="Message">
-/// The message to display.
-/// </param>
-public record SimpleWarning(string Message) : INotification;
