--- src/Ashampoo.Translation.Systems.Components/src/README.md
@@ -1,2 +1,0 @@
-# Ashampoo.Translation.Systems.Components
-Razor class library containing basic components and services for creating a converter with blazor.
