+++ src/Ashampoo.Translation.Systems.Logging/src/Notifications/FormatUploadedNotification.cs
@@ -1,0 +1,5 @@
+﻿using MediatR;
+
+namespace Ashampoo.Translation.Systems.Logging.Notifications;
+
+public record FormatUploadedNotification : INotification;
