+++ src/Ashampoo.Translation.Systems.Logging/src/Notifications/ImportedTranslationsNotification.cs
@@ -1,0 +1,5 @@
+﻿using MediatR;
+
+namespace Ashampoo.Translation.Systems.Logging.Notifications;
+
+public record ImportedTranslationsNotification(int Count) : INotification;
