+++ src/Ashampoo.Translation.Systems.Logging/src/Notifications/SimpleWarning.cs
@@ -1,0 +1,5 @@
+﻿using MediatR;
+
+namespace Ashampoo.Translation.Systems.Logging.Notifications;
+
+public record SimpleWarning(string Message) : INotification;
