+++ src/Ashampoo.Translation.Systems.Tools.ComponentLibrary/Ashampoo.Translation.Systems.Tools.ComponentLibrary.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk.Razor">
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="MediatR" Version="10.0.1" />
+        <PackageReference Include="MudBlazor" Version="6.0.14" />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\Ashampoo.Translation.Systems.Formats.Abstractions\src\Ashampoo.Translation.Systems.Formats.Abstractions.csproj" />
+        <ProjectReference Include="..\Ashampoo.Translation.Systems.Logging\src\Ashampoo.Translation.Systems.Logging.csproj" />
+    </ItemGroup>
+</Project>
