+++ src/Ashampoo.Translation.Systems.Tools.ComponentLibrary/Ashampoo.Translations.Tools.ComponentLibrary.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ashampoo.Translations.Tools.ComponentLibrary", "Ashampoo.Translations.Tools.ComponentLibrary.csproj", "{C1847618-47C8-4320-9249-3480E88BC18E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C1847618-47C8-4320-9249-3480E88BC18E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C1847618-47C8-4320-9249-3480E88BC18E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C1847618-47C8-4320-9249-3480E88BC18E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C1847618-47C8-4320-9249-3480E88BC18E}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
