+++ src/Ashampoo.Translation.Systems.Tools.ComponentLibrary/Components/DisplayFormat.razor
