+++ src/Ashampoo.Translation.Systems.Tools.ComponentLibrary/Dialogs/ConfigureFormatOptionsDialog.razor
@@ -1,0 +1,54 @@
+@using Ashampoo.Translation.Systems.Formats.Abstractions
+@using MudBlazor
+<MudDialog DefaultFocus="DefaultFocus.LastChild">
+    <DialogContent>
+        <MudForm @ref="form">
+
+            @foreach (var formatOption in FormatOptions.Options)
+            {
+                if (formatOption is FormatStringOption stringOption)
+                {
+                    <MudTextField T="string" Label="@formatOption.Name" AnchorOrigin="Origin.BottomCenter"
+                                  Required="@formatOption.Required" @bind-Value=@stringOption.Value ErrorText="@formatOption.Name">
+                    </MudTextField>
+                }
+
+                else if (formatOption is FormatFilterOption filterOption)
+                {
+                    <MudCheckBox @bind-Checked=filterOption.SetFilter Color="Color.Primary" Label="@formatOption.Name" Required="@formatOption.Required"></MudCheckBox>
+                }
+            }
+        </MudForm>
+    </DialogContent>
+
+    <DialogActions>
+        <MudButton OnClick="Cancel">Cancel</MudButton>
+        <MudButton Color="Color.Primary" OnClick="Submit">Ok</MudButton>
+    </DialogActions>
+
+</MudDialog>
+
+@code
+{
+    [CascadingParameter]
+    MudDialogInstance MudDialog { get; set; } = default!;
+
+
+    [Parameter]
+    public FormatOptions FormatOptions { get; init; } = default!;
+
+    private MudForm? form;
+
+    void Submit()
+    {
+        form?.Validate();
+        if (!form?.IsValid ?? true) return;
+        MudDialog.Close(DialogResult.Ok(true));
+    }
+
+    void Cancel()
+    {
+        form?.ResetValidation();
+        MudDialog.Cancel();
+    }
+}
