+++ src/Ashampoo.Translation.Systems.Tools.ComponentLibrary/Dialogs/SelectFormatDialog.razor
@@ -1,0 +1,41 @@
+@using Ashampoo.Translation.Systems.Formats.Abstractions
+@using MudBlazor
+@inject IFormatFactory FormatFactory
+
+<MudDialog DefaultFocus="DefaultFocus.LastChild">
+    <DialogContent>
+        <MudSelect T="string" Label="Select a Format" AnchorOrigin="Origin.BottomCenter" @bind-Value="@formatId" ErrorText="Please select a format" Error="@error">
+            @foreach (var provider in FormatFactory.GetFormatProviders())
+            {
+                <MudSelectItem Value="@provider.Id">@provider.Id</MudSelectItem>
+            }
+        </MudSelect>
+    </DialogContent>
+    <DialogActions>
+        <MudButton OnClick="Cancel">Cancel</MudButton>
+        <MudButton Color="Color.Primary" OnClick="Submit">Ok</MudButton>
+    </DialogActions>
+</MudDialog>
+
+@code
+{
+    [CascadingParameter]
+    MudDialogInstance MudDialog { get; set; } = default!;
+
+    private string? formatId;
+    private bool error;
+
+    void Submit()
+    {
+        if (string.IsNullOrWhiteSpace(formatId))
+        {
+            error = true;
+            return;
+        }
+        error = false;
+        MudDialog.Close(DialogResult.Ok(formatId));
+    }
+
+
+    void Cancel() => MudDialog.Cancel();
+}
