+++ src/Ashampoo.Translation.Systems.Tools.ComponentLibrary/NotificationHandlers/FormatUploadedNotificationHandler.cs
@@ -1,0 +1,22 @@
+﻿using Ashampoo.Translation.Systems.Logging.Notifications;
+using MediatR;
+using MudBlazor;
+
+namespace Ashampoo.Translation.Systems.Tools.ComponentLibrary.NotificationHandlers;
+
+public class FormatUploadedNotificationHandler : INotificationHandler<FormatUploadedNotification>
+{
+    private readonly ISnackbar snackbar;
+
+    public FormatUploadedNotificationHandler(ISnackbar snackbar)
+    {
+        this.snackbar = snackbar;
+    }
+
+    public Task Handle(FormatUploadedNotification notification, CancellationToken cancellationToken)
+    {
+        snackbar.Add("Uploaded Format.", Severity.Success);
+
+        return Task.CompletedTask;
+    }
+}
