+++ src/Ashampoo.Translation.Systems.Tools.ComponentLibrary/wwwroot/Scripts/WebFileService.js
